dtype float32

+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_basketball_v2.py
@@ -30,7 +30,7 @@ class SawyerBasketballEnvV2(SawyerXYZEnv):
             'obj_init_pos': np.array([0, 0.6, 0.03], dtype=np.float32),
             'hand_init_pos': np.array((0, 0.6, 0.2), dtype=np.float32),
         }
-        self.goal = np.array([0, 0.9, 0])
+        self.goal = np.array([0, 0.9, 0], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.hand_init_pos = self.init_config['hand_init_pos']
