+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_coffee_button_v2.py
@@ -28,11 +28,11 @@ class SawyerCoffeeButtonEnvV2(SawyerXYZEnv):
         )
 
         self.init_config = {
-            'obj_init_pos': np.array([0, 0.9, 0.28]),
+            'obj_init_pos': np.array([0, 0.9, 0.28], dtype=np.float32),
             'obj_init_angle': 0.3,
-            'hand_init_pos': np.array([0., .4, .2]),
+            'hand_init_pos': np.array([0., .4, .2], dtype=np.float32),
         }
-        self.goal = np.array([0, 0.78, 0.33])
+        self.goal = np.array([0, 0.78, 0.33], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.hand_init_pos = self.init_config['hand_init_pos']
