+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_door_close_v2.py
@@ -19,7 +19,7 @@ class SawyerDoorCloseEnvV2(SawyerDoorEnvV2):
             'obj_init_pos': np.array([0.1, 0.95, 0.15], dtype=np.float32),
             'hand_init_pos': np.array([-0.5, 0.6, 0.2], dtype=np.float32),
         }
-        self.goal = np.array([0.2, 0.8, 0.15])
+        self.goal = np.array([0.2, 0.8, 0.15], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.hand_init_pos = self.init_config['hand_init_pos']
