+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_handle_pull_side_v2.py
@@ -22,10 +22,10 @@ class SawyerHandlePullSideEnvV2(SawyerXYZEnv):
         )
 
         self.init_config = {
-            'obj_init_pos': np.array([-0.3, 0.7, 0.0]),
-            'hand_init_pos': np.array((0, 0.6, 0.2),),
+            'obj_init_pos': np.array([-0.3, 0.7, 0.0], dtype=np.float32),
+            'hand_init_pos': np.array((0, 0.6, 0.2), dtype=np.float32),
         }
-        self.goal = np.array([-0.2, 0.7, 0.14])
+        self.goal = np.array([-0.2, 0.7, 0.14], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.hand_init_pos = self.init_config['hand_init_pos']
