+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_pick_place_v2.py
@@ -36,11 +36,11 @@ class SawyerPickPlaceEnvV2(SawyerXYZEnv):
 
         self.init_config = {
             'obj_init_angle': .3,
-            'obj_init_pos': np.array([0, 0.6, 0.02]),
-            'hand_init_pos': np.array([0, .6, .2]),
+            'obj_init_pos': np.array([0, 0.6, 0.02], dtype=np.float32),
+            'hand_init_pos': np.array([0, .6, .2], dtype=np.float32),
         }
 
-        self.goal = np.array([0.1, 0.8, 0.2])
+        self.goal = np.array([0.1, 0.8, 0.2], dtype=np.float32)
 
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.obj_init_pos = self.init_config['obj_init_pos']
