+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_plate_slide_back_side_v2.py
@@ -41,7 +41,7 @@ class SawyerPlateSlideBackSideEnvV2(SawyerXYZEnv):
             'obj_init_pos': np.array([-0.25, 0.6, 0.02], dtype=np.float32),
             'hand_init_pos': np.array((0, 0.6, 0.2), dtype=np.float32),
         }
-        self.goal = np.array([0., 0.6, 0.015])
+        self.goal = np.array([0., 0.6, 0.015], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.hand_init_pos = self.init_config['hand_init_pos']
