+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_shelf_place_v2.py
@@ -24,7 +24,7 @@ class SawyerShelfPlaceEnvV2(SawyerXYZEnv):
         )
 
         self.init_config = {
-            'obj_init_pos':np.array([0, 0.6, 0.02]),
+            'obj_init_pos':np.array([0, 0.6, 0.02], dtype=np.float32),
             'obj_init_angle': 0.3,
             'hand_init_pos': np.array([0, 0.6, 0.2], dtype=np.float32),
         }
