+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_stick_pull_v2.py
@@ -24,8 +24,8 @@ class SawyerStickPullEnvV2(SawyerXYZEnv):
         )
 
         self.init_config = {
-            'stick_init_pos': np.array([0, 0.6, 0.02]),
-            'hand_init_pos': np.array([0, 0.6, 0.2]),
+            'stick_init_pos': np.array([0, 0.6, 0.02], dtype=np.float32),
+            'hand_init_pos': np.array([0, 0.6, 0.2], dtype=np.float32),
         }
         self.goal = self.init_config['stick_init_pos']
         self.stick_init_pos = self.init_config['stick_init_pos']
