+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_sweep_into_goal_v2.py
@@ -27,11 +27,11 @@ class SawyerSweepIntoGoalEnvV2(SawyerXYZEnv):
         )
 
         self.init_config = {
-            'obj_init_pos':np.array([0., 0.6, 0.02]),
+            'obj_init_pos':np.array([0., 0.6, 0.02], dtype=np.float32),
             'obj_init_angle': 0.3,
-            'hand_init_pos': np.array([0., .6, .2]),
+            'hand_init_pos': np.array([0., .6, .2], dtype=np.float32),
         }
-        self.goal = np.array([0., 0.84, 0.02])
+        self.goal = np.array([0., 0.84, 0.02], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.hand_init_pos = self.init_config['hand_init_pos']
